[chore] 함수명 변경 및 패키지 정리
