+++ app/src/main/java/com/hyeeyoung/wishboard/repository/cart/CartRepository.kt
@@ -1,6 +1,6 @@
 package com.hyeeyoung.wishboard.repository.cart
 
-import com.hyeeyoung.wishboard.model.CartItem
+import com.hyeeyoung.wishboard.model.cart.CartItem
 
 interface CartRepository {
     suspend fun addToCart(token: String, itemId: Int): Boolean
