--- app/src/main/java/com/hyeeyoung/wishboard/repository/home/WishRepository.kt
@@ -1,9 +1,0 @@
-package com.hyeeyoung.wishboard.repository.home
-
-import com.hyeeyoung.wishboard.model.WishItem
-import com.hyeeyoung.wishboard.model.WishItemInfo
-
-interface WishRepository {
-    suspend fun fetchItems(token: String): List<WishItem>?
-    suspend fun fetchItemDetail(token: String, itemId: Int): WishItemInfo?
-}
