+++ app/src/main/java/com/hyeeyoung/wishboard/repository/wish/WishRepository.kt
@@ -1,0 +1,9 @@
+package com.hyeeyoung.wishboard.repository.wish
+
+import com.hyeeyoung.wishboard.model.wish.WishItem
+import com.hyeeyoung.wishboard.model.wish.WishItemInfo
+
+interface WishRepository {
+    suspend fun fetchWishList(token: String): List<WishItem>?
+    suspend fun fetchWishItem(token: String, itemId: Int): WishItemInfo?
+}
