+++ app/src/main/java/com/hyeeyoung/wishboard/repository/wish/WishRepositoryImpl.kt
@@ -1,15 +1,15 @@
-package com.hyeeyoung.wishboard.repository.home
+package com.hyeeyoung.wishboard.repository.wish
 
 import android.util.Log
-import com.hyeeyoung.wishboard.model.WishItem
-import com.hyeeyoung.wishboard.model.WishItemInfo
+import com.hyeeyoung.wishboard.model.wish.WishItem
+import com.hyeeyoung.wishboard.model.wish.WishItemInfo
 import com.hyeeyoung.wishboard.remote.RemoteService
 
 class WishRepositoryImpl : WishRepository {
     private val api = RemoteService.api
 
-    override suspend fun fetchItems(token: String): List<WishItem>? {
-        val response = api.fetchItems(token) ?: return null
+    override suspend fun fetchWishList(token: String): List<WishItem>? {
+        val response = api.fetchWishList(token) ?: return null
 
         if (response.isSuccessful) {
             Log.d(TAG, "아이템 가져오기 성공")
@@ -19,8 +19,8 @@ class WishRepositoryImpl : WishRepository {
         return response.body()
     }
 
-    override suspend fun fetchItemDetail(token: String, itemId: Int): WishItemInfo? {
-        val response = api.fetchItemDetail(token, itemId) ?: return null
+    override suspend fun fetchWishItem(token: String, itemId: Int): WishItemInfo? {
+        val response = api.fetchWishItem(token, itemId) ?: return null
 
         if (response.isSuccessful) {
             Log.d(TAG, "아이템 상세정보 가져오기 성공")
